+++ src/main/java/cloud/metaapi/sdk/clients/meta_api/models/NewMetatraderAccountDto.java
@@ -66,8 +66,7 @@ public class NewMetatraderAccountDto {
      */
     public Map<String, Object> metadata;
     /**
-     * Used to increase the reliability of the account. Allowed values are regular, high
-     * and maximum. Default is regular
+     * Used to increase the reliability of the account. Allowed values are regular and high. Default is regular
      */
     public String reliability;
 }