+++ src/main/java/cloud/metaapi/sdk/clients/models/Error.java
@@ -1,5 +1,7 @@
 package cloud.metaapi.sdk.clients.models;
 
+import com.fasterxml.jackson.databind.JsonNode;
+
 /**
  * Contains an error message
  */
@@ -27,5 +29,9 @@ public class Error {
     /**
      * Additional information about error or {@code null}. Used to supply validation error details
      */
-    public Object details = null;
+    public JsonNode details = null;
+    /**
+     * Error metadata or {@code null}
+     */
+    public JsonNode metadata = null;
 }